Only update UI at 1 Hz during nandroid backup/restore
On high-resolution devices such as the Galaxy Note, UI updates
consume more CPU than all other operations.  For example, a nandroid
restore on Note pegs an entire core at 1.4 GHz for the recovery
process while tar is using only 14% of the other core, and iowait
times are low indicating the CPU is not simply waiting for IO.

Nandroid backup/restore should be I/O bound, not CPU bound.

This changes the update interval to 1 Hz during nandroid
backup and restore.

Change-Id: Ic021cd6dc453b92468396cab1a31bd693efb43e0

Signed-off-by: stratosk <stratosk@albert.(none)>