Use legacy property environment if old update-binary
If an installation package contains an update-binary that is
incompatible with the kitkit bionic property environment, set a legacy
property environment. This removes the need for a fallback update-binary
which is often incompatible with older update-scripts.

Method of setting prop environment based on cyanogen's commit:
"Set ANDROID_PROPERTY_WORKSPACE in exec." Tip to rename
/dev/__properties__ so that the legacy property workspace is actually
used comes from conversation with _that.

Change-Id: I5268f67537ad03c077fa05cbb98dfd4da194ffb8